nginx

+++ security/tests.php
@@ -296,34 +296,40 @@ function rsssl_code_execution_allowed()
 /**
  * Test if directory indexing is allowed
  * We assume allowed if test is not possible due to restrictions. Only an explicity 403 on the response results in "forbidden".
+ * On non htaccess servers, the default is non indexing, so we return forbidden.
+ *
  * @return bool
  */
 function rsssl_directory_indexing_allowed() {
+
 	$status = get_transient('rsssl_directory_indexing_status');
 	if ( !$status ) {
-		$status = 'allowed';
-
-		try {
-			$test_folder = 'indexing-test';
-			$test_dir = trailingslashit(ABSPATH) . $test_folder;
-			if ( ! is_dir( $test_dir ) ) {
-				mkdir( $test_dir, 0755 );
-			}
-
-			$response = wp_remote_get(trailingslashit( site_url($test_folder) ) );
-			if ( is_dir( $test_dir )  ) {
-				rmdir( $test_dir );
-			}
-
-			// WP_Error doesn't contain response code, return false
-			if ( !is_wp_error( $response ) ) {
-				$response_code = $response['response']['code'];
-				if ( $response_code === 403 ) {
-					$status = 'forbidden';
+		if ( !rsssl_uses_htaccess() ) {
+			$status = 'forbidden';
+		} else {
+			$status = 'allowed';
+			try {
+				$test_folder = 'indexing-test';
+				$test_dir = trailingslashit(ABSPATH) . $test_folder;
+				if ( ! is_dir( $test_dir ) ) {
+					mkdir( $test_dir, 0755 );
 				}
-			}
-		} catch( Exception $e ) {
 
+				$response = wp_remote_get(trailingslashit( site_url($test_folder) ) );
+				if ( is_dir( $test_dir )  ) {
+					rmdir( $test_dir );
+				}
+
+				// WP_Error doesn't contain response code, return false
+				if ( !is_wp_error( $response ) ) {
+					$response_code = $response['response']['code'];
+					if ( $response_code === 403 ) {
+						$status = 'forbidden';
+					}
+				}
+			} catch( Exception $e ) {
+
+			}
 		}
 
 		set_transient('rsssl_directory_indexing_status', $status, WEEK_IN_SECONDS );
